scheduling_automation/requirements_scheduler.txt
# Smile Dental Scheduling Automation — Dependencies

# Core
apscheduler>=3.10.0
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0

# Google Sheets (same as main app — shared token.pickle)
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0

# Env
python-dotenv>=1.0.0

# Timezone
tzdata
